Added monochrome output for some input formats.
For YUYV and UYUV input pixel formats, this will produce an 8-bit
monochrome output stream.  If you select an unsupported input format,
a warning will be printed instead.

I've also tweaked all the converters so that they do not use a temporary
buffer during conversion.  That saves a lot of time on some embedded
systems.